Struct vodozemac::ecies::EstablishedEcies

pub struct EstablishedEcies { /* private fields */ }

An established ECIES session.

This session can be used to encrypt and decrypt messages between the two sides of the channel.

Implementations

impl EstablishedEcies

pub fn public_key(&self) -> Curve25519PublicKey

Get our Curve25519PublicKey.

This public key needs to be sent to the other side so that it can complete the ECIES channel establishment.

pub fn check_code(&self) -> &CheckCode

Get the CheckCode which uniquely identifies this EstablishedEcies session.

This check code can be used to check that both sides of the session are indeed using the same shared secret.

pub fn encrypt(&mut self, plaintext: &[u8]) -> Message

Encrypt the given plaintext using this EstablishedEcies session.

pub fn decrypt(&mut self, message: &Message) -> Result<Vec<u8>, Error>

Decrypt the given message using this EstablishedEcies session.