vodozemac

Module hpke

Implementation of a hybrid public encryption scheme.

Examples

use vodozemac::hpke::*;

let plaintext = b"It's a secret to everybody";

let alice = HpkeSenderChannel::new();
let bob = HpkeRecipientChannel::new();

let SenderCreationResult { channel: mut alice, message } = alice
    .establish_channel(bob.public_key(), plaintext, &[])?;

let RecipientCreationResult { channel: mut bob, message } = bob.establish_channel(&message, &[])?;

assert_eq!(
    message, plaintext,
    "The decrypted plaintext should match our initial plaintext"
);

// Now we need to establish communication in the other direction, Bob
// needs to encrypt and send an initial reply to Alice.

let plaintext = b"Not a secret to me!";

let BidirectionalCreationResult { channel: mut bob, message } = bob.establish_bidirectional_channel(plaintext, &[]);
let BidirectionalCreationResult { channel: mut alice, message } = alice.establish_bidirectional_channel(&message, &[])?;

assert_eq!(message, plaintext);

// We now exchange the check code out-of-band and compare it.
if alice.check_code() != bob.check_code() {
    panic!("The check code must match; possible active MITM attack in progress");
}

let message = bob.seal(b"Another plaintext", &[]);
let decrypted = alice.open(&message, &[])?;

assert_eq!(decrypted, b"Another plaintext");

Structs

BidirectionalCreationResult

The result of the creation of a bidirectional and fully established HPKE channel.

CheckCode

A check code that can be used to confirm that two [EstablishedHpkeChannel] objects share the same secret. This is supposed to be shared out-of-band to protect against active MITM attacks.

EstablishedHpkeChannel

A fully established HPKE channel.

HpkeRecipientChannel

The unestablished HPKE recipient channel.

HpkeSenderChannel

The unestablished HPKE sender channel.

InitialMessage

The initial message, sent by the HPKE channel sender.

InitialResponse

The initial response, sent by the HPKE channel receiver.

Message

An encrypted message a EstablishedHpkeChannel channel has sent.

RecipientCreationResult

The result type for the initial establishment of a unidirectional HPKE channel.

SenderCreationResult

The result type for the initial establishment of a unidirectional HPKE channel.

UnidirectionalRecipientChannel

The unidirectional HPKE sender channel.

UnidirectionalSenderChannel

The unidirectional HPKE sender channel.

Enums

DigitMode

Enum controlling how the CheckCode should be represented as digits.

Error

The Error type for the HPKE submodule.

MessageDecodeError

The error type for the HPKE message decoding failures.