Module vodozemac::megolm

An implementation of the Megolm ratchet.

Structs

DecryptedMessage

ExportedSessionKey

The exported session key.

GroupSession

A Megolm group session represents a single sending participant in an encrypted group communication context containing multiple receiving parties.

GroupSessionPickle

A format suitable for serialization which implements serde::Serialize and serde::Deserialize. Obtainable by calling GroupSession::pickle.

InboundGroupSession

InboundGroupSessionPickle

A format suitable for serialization which implements serde::Serialize and serde::Deserialize. Obtainable by calling InboundGroupSession::pickle.

MegolmMessage

An encrypted Megolm message.

SessionKey

The session key, can be used to create a InboundGroupSession.

Enums

DecryptionError

Error type for Megolm-based decryption failuers.

SessionKeyDecodeError

Error type describing failure modes for the SessionKey and ExportedSessionKey decoding.