vodozemac

Module megolm

An implementation of the Megolm ratchet.

Structs

DecryptedMessage

A message successfully decrypted by an InboundGroupSession.

ExportedSessionKey

The exported session key.

GroupSession

A Megolm group session represents a single sending participant in an encrypted group communication context containing multiple receiving parties.

GroupSessionPickle

A format suitable for serialization which implements serde::Serialize and serde::Deserialize. Obtainable by calling GroupSession::pickle.

InboundGroupSession

A Megolm inbound group session represents a single receiving participant in an encrypted group communication involving multiple recipients.

InboundGroupSessionPickle

A format suitable for serialization which implements serde::Serialize and serde::Deserialize. Obtainable by calling InboundGroupSession::pickle.

MegolmMessage

An encrypted Megolm message.

SessionConfig

A struct to configure how Megolm sessions should work under the hood. Currently only the MAC truncation behaviour can be configured.

SessionKey

The session key, can be used to create a InboundGroupSession.

Enums

DecryptionError

Error type for Megolm-based decryption failures.

SessionKeyDecodeError

Error type describing failure modes for the SessionKey and ExportedSessionKey decoding.

SessionOrdering

The result of a comparison between two InboundGroupSession types.