Module vodozemac::olm

An implementation of the Olm ratchet.

Structs

Account

An Olm account manages all cryptographic keys used on a device.

AccountPickle

IdentityKeys

Struct holding the two public identity keys of an Account.

InboundCreationResult

Return type for the creation of inbound Session objects.

Message

An encrypted Olm message.

PreKeyMessage

An encrypted Olm pre-key message.

RatchetPublicKey

Session

An Olm session represents one end of an encrypted communication channel between two participants.

SessionKeys

The set of keys that were used to establish the Olm Session,

SessionPickle

A format suitable for serialization which implements serde::Serialize and serde::Deserialize. Obtainable by calling Session::pickle.

Enums

DecryptionError

MessageType

An enum over the two supported message types.

OlmMessage

Enum over the different Olm message types.

SessionCreationError